+++ QLDatLichKham/QLDatLichKham/src/dao/PhongKhamDAO.java
@@ -7,6 +7,6 @@ public interface PhongKhamDAO {
 	
 	List<PhongKham> findPhongKhamByDichVuIdAndNgayKhamAndBuoiKham(String dichVuId, Date ngayKham, int buoiKham);
 
-	List<PhongKham> findPhongKhamByLichBieu(Date ngayKham, int buoiKham, Long bacSiId);
+	List<PhongKham> findPhongKhamByLichBieu(Date ngayKham, int buoiKham, int bacSiId);
 
 }
